Update the Bot.ts framework into v9.0.0

+++ src/listeners/clean.guildMemberRemove.ts
@@ -1,30 +1,26 @@
-import * as app from "#app"
+import { Listener } from "#core/listener"
+import { getGuild, sendLog } from "#namespaces/tools"
+import users from "#tables/user"
 
-import users from "#tables/user.ts"
-
-export default new app.Listener({
+export default new Listener({
   event: "guildMemberRemove",
   description: "Delete member from db",
   async run(member) {
     const { guild } = member
 
-    const config = await app.getGuild(guild)
+    const config = await getGuild(guild)
 
-    try {
-      const user = await member.client.users.fetch(member.id)
+    const user = await member.client.users.fetch(member.id)
 
-      if (
-        member.client.guilds.cache
-          .filter((g) => g.id !== guild.id)
-          .every((g) => !g.members.cache.has(member.id))
-      )
-        throw new Error()
+    const isMemberInOtherGuilds = member.client.guilds.cache
+      .filter((g) => g.id !== guild.id)
+      .some((g) => g.members.cache.has(member.id))
 
-      await app.sendLog(guild, `${user} left the guild.`, config)
-    } catch {
+    if (isMemberInOtherGuilds) {
+      await sendLog(guild, `${user} left the guild.`, config)
+    } else {
       await users.query.delete().where({ id: member.id })
-
-      await app.sendLog(
+      await sendLog(
         guild,
         `${member.user ?? member} left all the labs.`,
         config,
